Server dev

See merge request s13-bigdata-recom-sub1/S13P21A103!12

+++ backend/src/main/java/com/ygss/backend/auth/dto/SignUpRequestDto.java
@@ -11,6 +11,6 @@ public class SignUpRequestDto {
     private String email;
     private String password;
     private Integer workedAt = 0;
-    private Integer salary;
-    private Integer totalRetirePension;
+    private Long salary;
+    private Long totalRetirePension;
 }
